Make the expose_unbound option actually do something
Now, if it's set and the zwave controller gets data from a device that
hasn't been configured with a nice name, you'll get an automatically
created item in the format 'zwave_{{ node_id }}_{{ label }}'. Thus far,
this has been sufficient to uniquely identify an item.

Also add a filter early in the notification handler to prevent operation
on values that fall outside of the supported types/genres.