Add home_id to automatic item/controller names
Items are now zwave_{{ home_id }}_{{ node_id }}_{{ label }}. This allows
for multiple zwave networks on the same bus.

Likewise, the controllers are now named
zwave_{{ home_id }}_Controller to allow for multiple
controllers on multiple networks. Currently, this doesn't differentiate
between multiple controllers on the same network as there is no
functionality that requires a specific controller. This is subject to
change in the future.

Also add home_id to zwave meta.